+++ src/report-category/report-category-resource.js
@@ -17,22 +17,22 @@
 
     'use strict';
 
-    /**
-     * @ngdoc object
-     * @name openlmis-rights.REPORT_RIGHTS
-     *
-     * @description
-     * This is constant for report rights.
-     */
     angular
-        .module('openlmis-rights')
-        .constant('REPORT_RIGHTS', rights());
+        .module('report-category')
+        .factory('ReportCategoryResource', ReportCategoryResource);
 
-    function rights() {
-        return {
-            REPORTS_VIEW: 'REPORTS_VIEW',
-            REPORT_TEMPLATES_EDIT: 'REPORT_TEMPLATES_EDIT'
-        };
+    ReportCategoryResource.inject = ['OpenlmisResource', 'classExtender'];
+
+    function ReportCategoryResource(OpenlmisResource, classExtender) {
+
+        classExtender.extend(ReportCategoryResource, OpenlmisResource);
+
+        return ReportCategoryResource;
+
+        function ReportCategoryResource() {
+            this.super('/api/reportCategory');
+        }
+
     }
 
 })();